Split React frontend from Flask backend
