+++ .env.example
@@ -5,4 +5,3 @@ RESUME_DIRECTORY_LOCATION=/tmp
 RESUME_FILENAME=Derek Santos - Resume.docx
 GENERATE_RESUME_ON_STARTUP=true
 GITHUB_TIMEOUT_SECONDS=5
-SPA_DIST_DIR=website/static/spa
