+++ README.md
@@ -1,4 +1,4 @@
-[![santosderek banner](website/static/images/santosderek.png)](https://santosderek.com)
+[![santosderek banner](frontend/public/static/images/santosderek.png)](https://santosderek.com)
 
 
 ![Build Passing](https://img.shields.io/github/actions/workflow/status/santosderek/website/tests.yml?branch=master&style=for-the-badge) ![Languages Count](https://img.shields.io/github/languages/count/santosderek/website?style=for-the-badge)
@@ -46,7 +46,7 @@ npm run build --prefix frontend
 npm test --prefix frontend
 ```
 
-The Flask app serves the built React Router app from `website/static/spa`. During frontend-only development, you can run Vite with API/static proxies to Flask:
+The React Router frontend is a standalone Vite app in `frontend/`. Flask is now API/download only. During local development, run Flask for `/api/v1/*` and `/resume`, and run Vite for the website UI with API proxies:
 
 ```bash
 # terminal 1
@@ -59,13 +59,11 @@ flask run --host 127.0.0.1 --port 8000
 npm run dev --prefix frontend
 ```
 
-To run the Flask development server locally with the production frontend build:
+To run the standalone frontend production preview:
 
 ```bash
-. .venv/bin/activate
-export FLASK_APP=website
-export FLASK_ENV=flask
-flask run --host 0.0.0.0 --port 8000
+npm run build --prefix frontend
+npm run preview --prefix frontend
 ```
 
 ## Dependency Workflow
@@ -92,8 +90,7 @@ Configuration can be supplied through environment variables or explicit app-fact
 | `RESUME_DIRECTORY_LOCATION` | `/tmp` on Linux, home directory otherwise | Directory where the generated resume is written. |
 | `RESUME_FILENAME` | `Derek Santos - Resume.docx` | Public download filename for `/resume`. |
 | `GENERATE_RESUME_ON_STARTUP` | `true` | Generates the DOCX resume when the app starts. Tests can disable this. |
-| `GITHUB_TIMEOUT_SECONDS` | `5` | Timeout for GitHub API requests used by the home page. |
-| `SPA_DIST_DIR` | `website/static/spa` | Directory containing the built React Router app served for `/` and known `/project/*` routes. |
+| `GITHUB_TIMEOUT_SECONDS` | `5` | Timeout for GitHub API requests used by the GitHub API endpoint. |
 
 ## Build
 
@@ -105,7 +102,7 @@ The workflow consists of:
 2. Running `python -m pip check`.
 3. Testing Python unit tests with PyTest.
 4. Installing and building the React Router frontend.
-5. Testing Dockerfile build of image.
+5. Testing Dockerfile build of the API image.
 6. Pushing the code to [DigitalOcean](https://www.digitalocean.com/). (Now handled by DigitalOcean's SaaS platform.)
 
 ## Docker
@@ -117,7 +114,7 @@ docker build -t santosderek-website .
 docker run --rm -p 8000:8000 santosderek-website
 ```
 
-The Docker image builds the React Router frontend in a Node stage, copies the generated `website/static/spa` files into the Python runtime image, and defaults to `FLASK_ENV=gunicorn`, which starts Gunicorn through `docker-entrypoint.sh`. Set `FLASK_ENV=flask` to run the Flask development server in the container.
+The Docker image is now for the Flask API/download backend only and defaults to `FLASK_ENV=gunicorn`, which starts Gunicorn through `docker-entrypoint.sh`. The React Router frontend is built and served separately with npm/Vite. Set `FLASK_ENV=flask` to run the Flask development server in the backend container.
 
 ## Deployment
