+++ frontend/public/sitemap.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
+  <url><loc>https://santosderek.com/</loc></url>
+  <url><loc>https://santosderek.com/project/project</loc></url>
+  <url><loc>https://santosderek.com/project/santosderek</loc></url>
+  <url><loc>https://santosderek.com/project/vitality</loc></url>
+</urlset>
