+++ frontend/public/static/images/vitality/DeploymentDiagram.png
