+++ frontend/vite.config.js
@@ -1,25 +1,21 @@
 import { defineConfig } from 'vite';
 import react from '@vitejs/plugin-react';
 
-const flaskOrigin = 'http://127.0.0.1:8000';
+const apiOrigin = 'http://127.0.0.1:8000';
 
 export default defineConfig({
   plugins: [react()],
-  base: '/static/spa/',
+  base: '/',
   build: {
-    outDir: '../website/static/spa',
+    outDir: 'dist',
     assetsDir: 'assets',
     emptyOutDir: true,
     manifest: true,
   },
   server: {
     proxy: {
-      '/api': flaskOrigin,
-      '/resume': flaskOrigin,
-      '/github': flaskOrigin,
-      '/linkedin': flaskOrigin,
-      '/robots.txt': flaskOrigin,
-      '/static': flaskOrigin,
+      '/api': apiOrigin,
+      '/resume': apiOrigin,
     },
   },
   test: {
