+++ requirements.in
@@ -1,5 +1,4 @@
 Flask>=2.2,<2.3
-flask-sitemap>=0.4,<0.5
 gunicorn>=23,<24
 python-docx>=1.1,<1.2
 python-dotenv>=1.0,<1.1
