+++ requirements.txt
@@ -3,7 +3,6 @@ certifi==2024.8.30
 charset-normalizer==3.4.0
 click==8.1.8
 Flask==2.2.5
-Flask-Sitemap==0.4.0
 gunicorn==23.0.0
 idna==3.10
 iniconfig==2.0.0
