--- website/sitemap/__init__.py
@@ -1,7 +1,0 @@
-from flask_sitemap import Sitemap
-
-
-sitemap = Sitemap()
-
-
-from . import website as website_sitemap
